requirements.txt
PyYAML>=5.1
sqlalchemy
requests
bs4
numpy
pandas
numexpr
flask
flask-caching
pytest
tox
coverage
Flask-SQLAlchemy
jinja2
geopy
records
click
flask_googlemaps
pylint
detect_delimiter
jinja2schema
pytest
pytest-bdd
sqlmodel